src/scraper/scraper.py
"""Web scraping module for property listings."""

import requests
from bs4 import BeautifulSoup
import logging
from typing import List, Dict

logger = logging.getLogger(__name__)


class PropertyScraper:
    """Base class for property scraping."""
    
    def __init__(self, timeout=10):
        """Initialize scraper with timeout."""
        self.timeout = timeout
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
    
    def fetch_page(self, url: str) -> BeautifulSoup:
        """Fetch and parse a web page."""
        try:
            response = requests.get(url, headers=self.headers, timeout=self.timeout)
            response.raise_for_status()
            return BeautifulSoup(response.content, 'html.parser')
        except requests.RequestException as e:
            logger.error(f"Error fetching {url}: {e}")
            raise
    
    def scrape_listings(self, location: str) -> List[Dict]:
        """Scrape property listings for a location."""
        raise NotImplementedError("Subclasses must implement scrape_listings")


class ZillowScraper(PropertyScraper):
    """Scraper for Zillow listings."""
    
    def scrape_listings(self, location: str) -> List[Dict]:
        """Scrape Zillow listings."""
        # Implementation would go here
        logger.info(f"Scraping Zillow listings for {location}")
        return []


class RedffinScraper(PropertyScraper):
    """Scraper for Redfin listings."""
    
    def scrape_listings(self, location: str) -> List[Dict]:
        """Scrape Redfin listings."""
        # Implementation would go here
        logger.info(f"Scraping Redfin listings for {location}")
        return []


def scrape_all_sources(location: str) -> List[Dict]:
    """Scrape listings from all configured sources."""
    all_listings = []
    
    scrapers = [
        ZillowScraper(),
        RedffinScraper(),
    ]
    
    for scraper in scrapers:
        try:
            listings = scraper.scrape_listings(location)
            all_listings.extend(listings)
        except Exception as e:
            logger.error(f"Error with {scraper.__class__.__name__}: {e}")
    
    return all_listings
