requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.11.2
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
APScheduler==3.10.1
python-dotenv==1.0.0
pytz==2023.3
lxml==4.9.3
